# apps/docs-platform/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VeriWorkly API",
    "version": "1.0.0",
...
  },
  "servers": [
    { "url": "https://api.veriworkly.com", "description": "Production Server" },
    { "url": "http://localhost:8080", "description": "Local Development" }
  ],
...
  "paths": {
...
    "/api/v1/documents": {
      "get": {
        "tags": ["Documents"],
        "operationId": "listDocuments",
        "security": [{ "ApiKeyAuth": [] }, { "LocalSession": [] }, { "ProdSession": [] }],
        "summary": "List Documents",
        "description": "Lists the authenticated user's documents (resumes, cover letters, portfolios, link-in-bio\npages), optionally filtered by type. Soft-deleted documents are excluded.\n",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["RESUME", "COVER_LETTER", "PORTFOLIO", "LINK_IN_BIO"]
            },
            "description": "Filter documents by type."
          }
        ],
        "responses": {
          "200": {
            "description": "Documents fetched successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DocumentListResponse" },
                "example": {
                  "success": true,
                  "message": "Documents fetched successfully",
                  "data": [
                    {
                      "id": "doc_123",
                      "userId": "user_123",
                      "type": "RESUME",
                      "title": "My Resume",
                      "slug": "my-resume",
                      "tags": [],
                      "content": {},
                      "metadata": {},
                      "templateId": "modern",
                      "schemaVersion": 1,
                      "revision": 1,
                      "visibility": "PRIVATE",
                      "lastSyncedAt": "2026-04-20T10:00:00.000Z",
                      "createdAt": "2026-04-01T10:00:00.000Z",
                      "updatedAt": "2026-04-20T10:00:00.000Z",
                      "deletedAt": null
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/UnauthorizedError" }
        }
      },
      "post": {
        "tags": ["Documents"],
        "operationId": "createDocument",
        "security": [{ "ApiKeyAuth": [] }, { "LocalSession": [] }, { "ProdSession": [] }],
        "summary": "Create Document",
        "description": "Creates a new document. If no `content` is provided for a RESUME or COVER_LETTER, it is\nauto-seeded from the user's MasterProfile (when one exists). Free-tier users are limited to\n1 active document per type; upgrading removes that limit.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DocumentCreateRequest" },
              "example": { "type": "RESUME", "title": "My Resume" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Document created successfully",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/DocumentResponse" } }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/UnauthorizedError" },
          "403": {
            "description": "Free-tier document limit reached",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "success": false,
                  "statusCode": 403,
                  "message": "Free users can only have 1 active resume at a time. Upgrade to Creator Pro for unlimited documents."
                }
              }
            }
          },
          "413": {
            "description": "Payload too large",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "success": false,
                  "statusCode": 413,
                  "message": "Document content payload is too large"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{id}": {
      "get": {
        "tags": ["Documents"],
        "operationId": "getDocument",
        "security": [{ "ApiKeyAuth": [] }, { "LocalSession": [] }, { "ProdSession": [] }],
        "summary": "Get Document",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Document fetched successfully",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/DocumentResponse" } }
            }
          },
          "401": { "$ref": "#/components/responses/UnauthorizedError" },
          "404": {
            "description": "Document not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "success": false, "statusCode": 404, "message": "Document not found" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Documents"],
        "operationId": "updateDocument",
        "security": [{ "ApiKeyAuth": [] }, { "LocalSession": [] }, { "ProdSession": [] }],
        "summary": "Update Document",
        "description": "Updates a document using optimistic concurrency control: `revision` must match the\ndocument's current revision or the request fails with `409`.\n",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DocumentUpdateRequest" },
              "example": { "title": "Updated Resume Title", "revision": 1 }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Document updated successfully",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/DocumentResponse" } }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/UnauthorizedError" },
          "404": {
            "description": "Document not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "success": false, "statusCode": 404, "message": "Document not found" }
              }
            }
          },
          "409": {
            "description": "Revision mismatch, or slug conflict",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "revisionMismatch": {
                    "summary": "Optimistic concurrency conflict",
                    "value": {
                      "success": false,
                      "statusCode": 409,
                      "message": "Revision mismatch. Client: 1, Server: 2"
                    }
                  },
                  "slugConflict": {
                    "summary": "Slug already in use",
                    "value": {
                      "success": false,
                      "statusCode": 409,
                      "message": "Document slug is already in use"
                    }
                  }
                }
              }
            }
          },
          "413": {
            "description": "Payload too large",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "success": false,
                  "statusCode": 413,
                  "message": "Document content payload is too large"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Documents"],
        "operationId": "deleteDocument",
        "security": [{ "ApiKeyAuth": [] }, { "LocalSession": [] }, { "ProdSession": [] }],
        "summary": "Delete Document",
        "description": "Soft-deletes a document by setting `deletedAt`. The document is immediately excluded from\nlist/get responses. Hard delete and restore exist at the service layer but are not exposed\nthrough any public route.\n\n`data` is `null` on success — the deleted document is not echoed back.\n",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Document deleted successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/NullDataResponse" },
                "example": {
                  "success": true,
                  "message": "Document deleted successfully",
                  "data": null
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/UnauthorizedError" },
          "404": {
            "description": "Document not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "success": false, "statusCode": 404, "message": "Document not found" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Your personal API key — `vw_` followed by 64 hex characters. The same key is also accepted\nas `Authorization: Bearer vw_...`.\n"
      },
      "LocalSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "veriworkly-auth.session_token",
        "description": "Internal session cookie for local development."
      },
      "ProdSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Secure-veriworkly-auth.session_token",
        "description": "Internal session cookie for production."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["success", "message", "statusCode"],
        "properties": {
          "success": { "type": "boolean" },
          "message": { "type": "string" },
          "statusCode": { "type": "integer" },
          "details": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "object",
              "properties": { "path": { "type": "string" }, "message": { "type": "string" } }
            }
          }
        }
      },
...
      "NullDataResponse": {
        "type": "object",
        "description": "Success envelope for endpoints that acknowledge an action without returning a body.",
        "required": ["success", "message", "data"],
        "properties": {
          "success": { "type": "boolean", "const": true },
          "message": { "type": "string" },
          "data": { "type": "null" }
        }
      },
...
      "Document": {
        "type": "object",
        "required": [
          "id",
          "userId",
          "type",
          "title",
          "slug",
          "tags",
          "content",
          "templateId",
          "schemaVersion",
          "revision",
          "visibility",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "type": {
            "type": "string",
            "enum": ["RESUME", "COVER_LETTER", "PORTFOLIO", "LINK_IN_BIO"]
          },
          "title": { "type": "string" },
          "slug": {
            "type": "string",
            "description": "Unique per user. Also used as the portfolio subdomain slug for PORTFOLIO documents."
          },
          "tags": { "type": "array", "items": { "type": "string" } },
          "content": {
            "type": "object",
            "additionalProperties": true,
            "description": "Document-type-specific JSON payload (resume schema, portfolio schema, etc)."
          },
          "metadata": { "type": "object", "nullable": true, "additionalProperties": true },
          "templateId": { "type": "string", "default": "modern" },
          "schemaVersion": { "type": "integer" },
          "revision": {
            "type": "integer",
            "description": "Incremented on every update. Required (and checked) on PATCH for optimistic concurrency control."
          },
          "visibility": { "type": "string", "enum": ["PRIVATE", "UNLISTED", "PUBLIC"] },
          "lastSyncedAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Set by soft delete. Documents with a non-null deletedAt are excluded from list/get."
          }
        }
      },
      "DocumentListResponse": {
        "type": "object",
        "required": ["success", "message", "data"],
        "properties": {
          "success": { "type": "boolean" },
          "message": { "type": "string" },
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/Document" } }
        }
      },
      "DocumentResponse": {
        "type": "object",
        "required": ["success", "message", "data"],
        "properties": {
          "success": { "type": "boolean" },
          "message": { "type": "string" },
          "data": { "$ref": "#/components/schemas/Document" }
        }
      },
      "DocumentCreateRequest": {
        "type": "object",
        "required": ["type"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Optional client-generated id. If a document with this id already exists for the caller, it is returned as-is (idempotent create)."
          },
          "type": {
            "type": "string",
            "enum": ["RESUME", "COVER_LETTER", "PORTFOLIO", "LINK_IN_BIO"]
          },
          "title": { "type": "string" },
          "slug": { "type": "string" },
          "tags": { "type": "array", "items": { "type": "string" } },
          "content": {
            "type": "object",
            "additionalProperties": true,
            "description": "If omitted for RESUME/COVER_LETTER, content is auto-seeded from the user's MasterProfile when one exists."
          },
          "metadata": { "type": "object", "additionalProperties": true },
          "templateId": { "type": "string" },
          "visibility": { "type": "string", "enum": ["PRIVATE", "UNLISTED", "PUBLIC"] }
        }
      },
      "DocumentUpdateRequest": {
        "type": "object",
        "required": ["revision"],
        "properties": {
          "title": { "type": "string" },
          "slug": { "type": "string" },
          "updateShareSlug": {
            "type": "boolean",
            "description": "When true and an active share link exists, its slug is updated to track the new document slug."
          },
          "tags": { "type": "array", "items": { "type": "string" } },
          "content": { "type": "object", "additionalProperties": true },
          "metadata": { "type": "object", "additionalProperties": true },
          "templateId": { "type": "string" },
          "visibility": { "type": "string", "enum": ["PRIVATE", "UNLISTED", "PUBLIC"] },
          "revision": {
            "type": "integer",
            "description": "Must match the document's current revision. Mismatch returns 409."
          }
        }
      },
...
    },
    "responses": {
      "ValidationError": {
        "description": "Bad Request - The request body or parameters are invalid.",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
        }
      },
      "UnauthorizedError": {
        "description": "Unauthorized - Authentication failed or session expired.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "success": false, "statusCode": 401, "message": "Authentication required" }
          }
        }
      },
...
    }
  }
}